requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
httpx==0.25.1
python-multipart==0.0.6
pydantic-settings==2.1.0
azure-core
flyr
openpyxl
requests
numpy
pandas
opencv-python
pythonnet
black
aiortc
pillow
psutil
python-dotenv
channels>=4.0.0
redis